requirements.txt
asgiref
async-timeout
attrs
autobahn
Automat
channels
constantly
daphne
dj-database-url
Django
django-heroku
djangorestframework
gunicorn
hyperlink
idna
incremental
lxml
mysqlclient
Pillow
psycopg2
PyHamcrest
pypng
python-pptx
pytz
qrcode
six
Twisted
txaio
uuid
whitenoise
XlsxWriter
zope.interface
